Move implementation to its own class
