+++ libs/natls/src/main/java/org/amshove/natls/languageserver/NaturalLanguageService.java
@@ -18,15 +18,22 @@ import org.amshove.natls.hover.HoverProvider;
 import org.amshove.natls.inlayhints.InlayHintProvider;
 import org.amshove.natls.progress.IProgressMonitor;
 import org.amshove.natls.progress.ProgressTasks;
-import org.amshove.natls.project.*;
+import org.amshove.natls.project.LanguageServerFile;
+import org.amshove.natls.project.LanguageServerProject;
+import org.amshove.natls.project.ModuleReferenceParser;
+import org.amshove.natls.project.ParseStrategy;
 import org.amshove.natls.referencing.ReferenceFinder;
 import org.amshove.natls.signaturehelp.SignatureHelpProvider;
 import org.amshove.natls.snippets.SnippetEngine;
+import org.amshove.natls.workspace.RenameFileHandler;
 import org.amshove.natparse.NodeUtil;
 import org.amshove.natparse.infrastructure.ActualFilesystem;
 import org.amshove.natparse.lexing.SyntaxKind;
 import org.amshove.natparse.lexing.SyntaxToken;
-import org.amshove.natparse.natural.*;
+import org.amshove.natparse.natural.IModuleReferencingNode;
+import org.amshove.natparse.natural.IReferencableNode;
+import org.amshove.natparse.natural.ISymbolReferenceNode;
+import org.amshove.natparse.natural.IVariableReferenceNode;
 import org.amshove.natparse.natural.project.NaturalFile;
 import org.amshove.natparse.natural.project.NaturalFileType;
 import org.amshove.natparse.natural.project.NaturalProject;
@@ -44,7 +51,6 @@ import java.io.UncheckedIOException;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.util.ArrayList;
-import java.util.HashMap;
 import java.util.List;
 import java.util.UUID;
 import java.util.concurrent.CompletableFuture;
@@ -62,6 +68,7 @@ public class NaturalLanguageService implements LanguageClientAware
 	private final InlayHintProvider inlayHintProvider = new InlayHintProvider();
 	private HoverProvider hoverProvider;
 	private final RenameSymbolAction renameComputer = new RenameSymbolAction();
+	private final RenameFileHandler renameFileHandler = new RenameFileHandler();
 	private final CodeLensService codeLensService = new CodeLensService();
 
 	private static LSConfiguration config = LSConfiguration.createDefault();
@@ -682,94 +689,7 @@ public class NaturalLanguageService implements LanguageClientAware
 
 	public WorkspaceEdit willRenameFiles(List<FileRename> renames)
 	{
-		var edit = new WorkspaceEdit();
-		var fileChanges = new HashMap<String, List<TextEdit>>();
-
-		for (var rename : renames)
-		{
-			var newNameRaw = LspUtil.uriToPath(rename.getNewUri()).getFileName().toString().split("\\.")[0];
-
-			var oldPath = LspUtil.uriToPath(rename.getOldUri());
-			var oldFile = findNaturalFile(oldPath);
-
-			if (oldFile.getType() == NaturalFileType.SUBROUTINE)
-			{
-				// Rename of file doesn't change the referable name of external subroutines
-				continue;
-			}
-
-			if (newNameRaw.toUpperCase().equals(oldFile.getReferableName()))
-			{
-				// The file has just been moved, the name remains the same
-				continue;
-			}
-
-			// TODO: Do name clash check if a module with new referable name in lib already exists
-
-			oldFile.parse(ParseStrategy.WITH_CALLERS);
-			var oldModule = oldFile.module();
-
-			var newName = switch (oldModule.file().getFiletype())
-			{
-				case SUBPROGRAM, PROGRAM -> "'%s'".formatted(newNameRaw);
-				default -> newNameRaw;
-			};
-
-			var callers = oldModule.callers();
-
-			for (var caller : callers)
-			{
-				var changes = fileChanges.computeIfAbsent(caller.referencingToken().filePath().toUri().toString(), u -> new ArrayList<>());
-				var textEdit = new TextEdit();
-				textEdit.setNewText(newName);
-				textEdit.setRange(LspUtil.toRange(caller.referencingToken()));
-				changes.add(textEdit);
-			}
-
-			for (var cachedPosition : ModuleReferenceCache.retrieveCachedPositions(oldFile))
-			{
-				var changes = fileChanges.computeIfAbsent(cachedPosition.filePath().toUri().toString(), u -> new ArrayList<>());
-				var textEdit = new TextEdit();
-				textEdit.setNewText(newName);
-				textEdit.setRange(LspUtil.toRange(cachedPosition));
-				changes.add(textEdit);
-			}
-
-			if (oldFile.getType() == NaturalFileType.FUNCTION && oldModule instanceof IFunction func)
-			{
-				var changes = fileChanges.computeIfAbsent(oldModule.file().getPath().toUri().toString(), u -> new ArrayList<>());
-
-				var textEdit = new TextEdit();
-				textEdit.setNewText(newName);
-				textEdit.setRange(LspUtil.toRange(func.functionName()));
-				changes.add(textEdit);
-
-				var funcNameVariable = func.defineData().findVariable(oldFile.getReferableName());
-				if (funcNameVariable != null)
-				{
-					for (var reference : funcNameVariable.references())
-					{
-						var refEdit = new TextEdit();
-						refEdit.setNewText(newName);
-						refEdit.setRange(LspUtil.toRange(reference.referencingToken()));
-						changes.add(refEdit);
-					}
-				}
-			}
-
-			// This is here and not in didRename because the client changes the files after this call.
-			// The changed files then can't reference the new file if `renameFile` wasn't run.
-			// Moving this to `didRename`, which sounds fine, can't take up the references to the file anymore,
-			// because they're gone.
-			// They're gone because the changes in the referencing files cause a reparse.
-
-			// TODO: Unless we can serialize the WorkspaceEdit and do it after didRename?
-			languageServerProject.renameFile(rename.getOldUri(), rename.getNewUri());
-		}
-
-		edit.setChanges(fileChanges);
-
-		return edit;
+		return renameFileHandler.handleFileRename(renames, languageServerProject);
 	}
 
 	private static <T> T extractJsonObject(Object obj, Class<T> clazz)
